requirements.txt
fastapi
uvicorn
boto3
scikit-learn
numpy
pandas
python-multipart
jinja2
python-dotenv
SQLAlchemy
cryptography
pymysql
aiomysql